split checkbox renderer file
